pyproject.toml
[project]
name = "dio-document-fraude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-core>=1.38.2",
    "azure-storage-blob>=12.28.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.54.0",
]
